Just copy&paste this hook to your faction file.
This is simply an HL2TS2 template for those who are lazy to add custom models to their citizen faction. This model pack includes HD versions of the original Citizen models and a few other additions from different games. You must download HL2TS2 Models and HL2TS2 Content from the workshop for it to work. 

Feel free to add any bodygroups, all credits go to the creators of this model pack. Make sure to set the ix.anim.SetModelClass in sh_schema.lua accordingly to your models so the animations work properly.